+++ lib/elasticsearch_record/gem_version.rb
@@ -9,7 +9,7 @@ module ElasticsearchRecord
   module VERSION
     MAJOR = 1
     MINOR = 7
-    TINY  = 6
+    TINY  = 7
     PRE   = nil
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")